Refactoring, cleanup, jtr module, suid module

+++ modules/P11_dlink_SHRS_enc_extract.sh
@@ -3,7 +3,6 @@
 # EMBA - EMBEDDED LINUX ANALYZER
 #
 # Copyright 2020-2022 Siemens Energy AG
-# Copyright 2020-2022 Siemens AG
 #
 # EMBA comes with ABSOLUTELY NO WARRANTY. This is free software, and you are
 # welcome to redistribute it under the terms of the GNU General Public License.
@@ -11,7 +10,7 @@
 #
 # EMBA is licensed under GPLv3
 #
-# Author(s): Michael Messner, Pascal Eckmann
+# Author(s): Michael Messner
 # Contributor: Benedikt Kuehne
 
 # Description: Extracts encrypted firmware images from D-Link
@@ -21,7 +20,7 @@ export PRE_THREAD_ENA=0
 
 P11_dlink_SHRS_enc_extract() {
   module_log_init "${FUNCNAME[0]}"
-  NEG_LOG=0
+  local NEG_LOG=0
 
   if [[ "$DLINK_ENC_DETECTED" -ne 0 ]]; then
     module_title "DLink encrypted firmware extractor"
@@ -40,20 +39,25 @@ P11_dlink_SHRS_enc_extract() {
 }
 
 dlink_SHRS_enc_extractor() {
-  local DLINK_ENC_PATH_="$1"
-  local EXTRACTION_FILE_="$2"
+  local DLINK_ENC_PATH_="${1:-}"
+  local EXTRACTION_FILE_="${2:-}"
+  if ! [[ -f "$DLINK_ENC_PATH_" ]]; then
+    print_output "[-] No file for decryption provided"
+    return
+  fi
+
   sub_module_title "DLink encrypted firmware extractor"
 
   hexdump -C "$DLINK_ENC_PATH_" | head | tee -a "$LOG_FILE" || true
 
-  print_output ""
+  print_ln
 
   dd if="$DLINK_ENC_PATH_" skip=1756 iflag=skip_bytes|openssl aes-128-cbc -d -p -nopad -nosalt -K "c05fbf1936c99429ce2a0781f08d6ad8" -iv "67c6697351ff4aec29cdbaabf2fbe346" --nosalt -in /dev/stdin -out "$EXTRACTION_FILE_" 2>&1 || true | tee -a "$LOG_FILE"
 
-  print_output ""
+  print_ln
   if [[ -f "$EXTRACTION_FILE_" ]]; then
     print_output "[+] Decrypted D-Link firmware file to $ORANGE$EXTRACTION_FILE_$NC"
-    print_output ""
+    print_ln
     print_output "[*] Firmware file details: $ORANGE$(file "$EXTRACTION_FILE_")$NC"
     write_csv_log "Extractor module" "Original file" "extracted file/dir" "file counter" "directory counter" "further details"
     write_csv_log "DLink SHRS decryptor" "$DLINK_ENC_PATH_" "$EXTRACTION_FILE_" "1" "NA" "NA"
@@ -68,9 +72,16 @@ dlink_SHRS_enc_extractor() {
 
 dlink_enc_img_extractor(){
   local TMP_DIR="$LOG_DIR""/tmp"
-  local DLINK_ENC_PATH_="$1"
-  local EXTRACTION_FILE_="$2"
+  local DLINK_ENC_PATH_="${1:-}"
+  local EXTRACTION_FILE_="${2:-}"
   local TMP_IMAGE_FILE="$TMP_DIR/image.bin"
+  if ! [[ -f "$DLINK_ENC_PATH_" ]]; then
+    print_output "[-] No file for decryption provided"
+    return
+  fi
+  local IMAGE_SIZE=0
+  local OFFSET=0
+  local ITERATION=0
 
   sub_module_title "DLink encrpted_image extractor"
 
@@ -90,10 +101,10 @@ dlink_enc_img_extractor(){
     --nosalt | dd if=/dev/stdin of="$EXTRACTION_FILE_" oflag=append conv=notrunc 2>&1 | tee -a "$LOG_FILE"
   done
   # Now it should be a .ubi file thats somewhat readable and extractable via ubireader
-  print_output ""
+  print_ln
   if [[ -f "$EXTRACTION_FILE_" ]]; then
     print_output "[+] Decrypted D-Link firmware file to $ORANGE$EXTRACTION_FILE_$NC"
-    print_output ""
+    print_ln
     print_output "[*] Firmware file details: $ORANGE$(file "$EXTRACTION_FILE_")$NC"
     write_csv_log "Extractor module" "Original file" "extracted file/dir" "file counter" "directory counter" "further details"
     write_csv_log "DLink enc_img decryptor" "$DLINK_ENC_PATH_" "$EXTRACTION_FILE_" "1" "NA" "NA"
